chore: use templates as io_uring operations wrapper
- Using templates allows us to make io_uring submission operation
more generic. With this design, adding a new operation is as
simple as calling the helper methods with the args of the new
operation.
- Use compile time check to verify if we can want to use async
submission or not. In this context, async submission is an
internal io_uring machinery. It does not mean our code is no
longer async. Please check IOSQE_ASYNC flag manual for more info.